backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart

# Environment & Config
python-dotenv
pydantic
pydantic-settings

# Google Gemini SDK (new client API)
google-genai>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0
aiosqlite

# HTTP Client
httpx==0.26.0

# LangGraph Agent System
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
